fix circular dependencies
Moved `getIconUrl` to Icons/index.ts since this caused most of the circular dependencies (this is BREAKING CHANGE). Replaced problematic barrel imports with relative imports.

- fix some of the linting/typescript errors/warnings along the way
