+++ src/components/common/Icons/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import Identicon from '@polkadot/react-identicon';
-import { getIconUrl } from '../../../utils/utils';
+import { reefTokenWithAmount } from '../../../state/token';
+import { getHashSumLastNr } from '../../../utils/math';
 
-export const DEFAULT_TOKEN_ICONS: string[] = [
+const DEFAULT_TOKEN_ICONS: string[] = [
   '<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="300" height="300" fill="none"><style>.B{fill-rule:evenodd}.C{fill:#fff}.D{fill:#e040fb}.E{fill:#d500f9}.F{fill:#651fff}</style><rect width="300" height="300" rx="150" fill="#fffbf8"/><g class="B"><g fill="#311b92"><path d="M158.233 139.522h-15.685c-6.4 0-11.636-5.238-11.636-11.637v-1.603c0-6.399 5.236-11.636 11.636-11.636h15.685c6.399 0 11.635 5.237 11.635 11.636v1.603c0 6.399-5.236 11.637-11.635 11.637"/><path d="M152.52 121.394c1.321 1.005 2.107 2.18 2.644 3.204.531 1.036.8 1.932 1.025 2.783.423 1.683.522 3.113.588 4.479.11 2.717-.074 5.179-.307 7.601-.473 4.819-1.283 9.413-2.083 13.983l-5.269 27.011a555.74 555.74 0 0 0-4.5 26.908c-1.266 9.004-2.164 18.043-2.561 27.267a.41.41 0 0 1-.426.394.412.412 0 0 1-.384-.316c-2.114-9.072-2.743-18.461-2.748-27.77-.009-9.329.803-18.609 1.819-27.814 1.026-9.204 2.398-18.342 3.67-27.359.652-4.495 1.238-8.986 1.639-13.309.181-2.15.291-4.281.232-6.153-.027-.926-.135-1.796-.256-2.345-.057-.26-.127-.469-.121-.456 0-.03.013.032.03.046.019.02.039.032.085.108.045.066.089.084.161.166a1 1 0 0 0 .253.205 5.41 5.41 0 0 1-1.062-7.572 5.406 5.406 0 0 1 7.571-1.061"/><path d="M154.236 130.305a.98.98 0 0 0 .253-.205c.072-.084.116-.1.16-.167.047-.076.067-.089.086-.108.017-.014.03-.076.029-.045.007-.012-.063.195-.121.456-.12.548-.228 1.42-.255 2.344-.059 1.874.052 4.005.232 6.153.401 4.325.987 8.814 1.639 13.31l3.67 27.358c1.016 9.207 1.828 18.487 1.82 27.814-.007 9.309-.635 18.7-2.75 27.77a.41.41 0 0 1-.809-.077c-.396-9.225-1.296-18.264-2.561-27.269a553.633 553.633 0 0 0-4.501-26.906l-5.268-27.011c-.801-4.569-1.61-9.163-2.083-13.982-.233-2.423-.416-4.884-.307-7.602.066-1.367.165-2.796.588-4.479.225-.851.493-1.748 1.025-2.784.537-1.025 1.323-2.199 2.645-3.203a5.405 5.405 0 0 1 7.57 1.061 5.41 5.41 0 0 1-1.062 7.572"/></g><path d="M182.452 231.463c-1.039-12.672-2.998-25.267-5.46-37.734-2.454-12.47-5.461-24.821-8.473-37.042l-4.231-18.144a115.02 115.02 0 0 1-1.105-6.16c2.078-1.76 4.025-3.877 5.801-6.25 6.739-9.006 11.05-26.937 11.05-36.8C180.034 72.029 166.762 58 150.39 58s-29.645 14.029-29.645 31.333c0 12.514 6.943 32.206 16.976 42.156-.301 2.263-.739 4.651-1.228 7.054-1.232 5.951-2.74 12.039-4.233 18.144-3.011 12.221-6.019 24.572-8.472 37.042-2.461 12.467-4.421 25.062-5.46 37.734-1.035 12.663-1.155 25.42.205 37.992a.615.615 0 0 0 .613.545.61.61 0 0 0 .605-.617c-.105-12.547 1.264-25.012 3.504-37.272 2.248-12.267 5.378-24.342 8.993-36.274 3.612-11.942 7.749-23.713 11.958-35.579l6.171-18.116.013-.042.013.042c1.949 6.215 4.073 12.169 6.17 18.116 4.209 11.866 8.348 23.637 11.958 35.579 3.617 11.932 6.747 24.007 8.994 36.274 2.241 12.26 3.61 24.725 3.505 37.272a.614.614 0 0 0 .544.613.61.61 0 0 0 .673-.541c1.361-12.572 1.24-25.329.205-37.992" class="D"/></g><mask id="A" maskUnits="userSpaceOnUse" x="117" y="58" width="67" height="212" mask-type="alpha"><path d="M120.745 89.334c0 12.514 6.944 32.204 16.976 42.156-.302 2.262-.739 4.65-1.228 7.052-1.233 5.953-2.74 12.041-4.232 18.144-3.012 12.221-6.02 24.574-8.472 37.042-2.462 12.467-4.421 25.063-5.46 37.735-1.035 12.662-1.157 25.419.205 37.991a.61.61 0 0 0 .612.545.61.61 0 0 0 .604-.616c-.104-12.547 1.264-25.011 3.505-37.272 2.247-12.266 5.378-24.341 8.993-36.273 3.611-11.943 7.749-23.713 11.959-35.581l6.17-18.115c.005-.014.008-.028.014-.042a.33.33 0 0 0 .012.042c1.949 6.215 4.074 12.169 6.17 18.115 4.21 11.868 8.348 23.638 11.958 35.581 3.617 11.932 6.747 24.007 8.993 36.273 2.242 12.261 3.611 24.725 3.506 37.272a.61.61 0 0 0 .545.612.61.61 0 0 0 .672-.541c1.36-12.572 1.239-25.329.205-37.991-1.04-12.672-2.999-25.268-5.461-37.735-2.453-12.468-5.461-24.821-8.472-37.042l-4.232-18.144a115.779 115.779 0 0 1-1.105-6.158c2.079-1.76 4.026-3.877 5.802-6.25 6.739-9.007 11.05-26.939 11.05-36.8 0-17.306-13.272-31.333-29.643-31.333-16.373 0-29.646 14.028-29.646 31.333" class="B C"/></mask><g mask="url(#A)" class="B E"><path d="m182.835 267.613-1.247-4.157-1.578-4.059c-.492-1.355-.486-2.875-.998-4.204l-1.655-3.908c-.501-1.302-1.424-2.479-1.849-3.711-.321-.944.275-1.146-.037-2.094l-.523-1.915c-.359-1.27-.993-2.466-1.27-3.746-.28-1.286-.966-2.53-1.136-3.833-.149-1.06.659-2.227.582-3.296-.078-1.062.105-2.112.097-3.186l.095-3.174.122-3.182c.112-1.051.542-2.048.71-3.119.211-1.389-.334-2.93-.04-4.368l.962-4.401c.373-1.564.727-1.478 1.103-3.042s.769-1.471 1.146-3.036l.63-3.162c.259-1.111-.44-2.457-.197-3.604l.796-3.376.632-3.41c.117-.887 1.112-1.688 1.197-2.585.082-.892-.544-1.853-.509-2.749.045-.897.411-1.799.402-2.695s-.492-1.789-.555-2.682c-.114-1.783.466-3.682.132-5.391s-1.937-3.07-2.488-4.656c-.563-1.581-1.13-3.074-1.889-4.5-.369-.724.066-1.901-.34-2.586-.43-.666-.982-1.297-1.439-1.93-.343-.474-.552-.329-.903-.797s-.08-.678-.437-1.137c-.469-.616-1.399-.864-1.902-1.443-.317-.353-.079-.578-.403-.922-.216-.221-.375-.495-.587-.725-.425-.458-1.683-.203-2.091-.688-.819-.99-1.091-2.458-1.872-3.522-.761-1.039-.958-2.472-1.682-3.57l-2.29-3.307c-.726-1.149-2.324-1.789-3.02-2.984-.688-1.204-.919-2.693-1.601-3.93-.673-1.222-1.021-2.633-1.693-3.867-.008-.013-.06.015-.068.001-.264-.487-.846.043-1.443.447-.379.257-1.431.091-1.492.464-.265 1.703.785 3.419.708 5.053-.027.81-.559 1.588-.566 2.395.005.801.503 1.577.528 2.375.057 1.596-.706 3.215-.541 4.785.181 1.561.784 3.059 1.084 4.588l.48 3.076c.255 1.004.661 1.967.975 2.953s1.084 1.79 1.461 2.756l.971 2.093c.159.356-.758 1.234-.578 1.582.261.512.921.18 1.173.697.542 1.147.589 2.477 1.033 3.64.462 1.154 1.41 2.116 1.752 3.289s.068 2.466.272 3.664a31.42 31.42 0 0 1 .481 3.598c.092 1.208.821 2.401.799 3.646-.016 1.104-.568 2.201-.677 3.355l-.217 3.377c-.165 1.09-.986 2.082-1.202 3.219-.206 1.086-.128 2.241-.37 3.372l-.735 3.309c-.238 1.056.317 2.309.072 3.399-.238 1.063-.874 2.052-1.107 3.144-.227 1.068.174 2.27-.035 3.364-.326 1.681-1.699 3.203-1.949 4.932-.24 1.727.283 3.554.152 5.304l-.314 3.448c-.015 1.161.198 2.302.253 3.451.056 1.163-.419 2.364-.278 3.503.201 1.689 1.067 3.245 1.452 4.851.368 1.612.51 3.255 1.044 4.765.348.99.462 2.076.878 3.018.422.959 1.384 1.65 1.866 2.56.491.924.783 1.945 1.324 2.821.399.646.674 1.375 1.089 2.002.439.61 1.525.719 1.987 1.312.234.294.39.632.625.914.353.418.423.354.778.77.471.548.556 1.41 1.041 1.933s.81 1.191 1.304 1.697c.486.513.859 1.145 1.364 1.64l3.015 2.994c1.042.964 2.772 1.144 3.858 2.07 1.076.916 1.391 2.793 2.54 3.652.663.495 1.574-.702 2.291-1.508.492-.554.89-.166.737-.693-.233-.133-.189-.424-.19-.428"/></g><mask id="B" maskUnits="userSpaceOnUse" x="117" y="58" width="67" height="212" mask-type="alpha"><path d="M120.745 89.334c0 12.514 6.944 32.204 16.976 42.156-.302 2.262-.739 4.65-1.228 7.052-1.233 5.953-2.74 12.041-4.232 18.144-3.012 12.221-6.02 24.574-8.472 37.042-2.462 12.467-4.421 25.063-5.46 37.735-1.035 12.662-1.157 25.419.205 37.991a.61.61 0 0 0 .612.545.61.61 0 0 0 .604-.616c-.104-12.547 1.264-25.011 3.505-37.272 2.247-12.266 5.378-24.341 8.993-36.273 3.611-11.943 7.749-23.713 11.959-35.581l6.17-18.115c.005-.014.008-.028.014-.042a.33.33 0 0 0 .012.042c1.949 6.215 4.074 12.169 6.17 18.115 4.21 11.868 8.348 23.638 11.958 35.581 3.617 11.932 6.747 24.007 8.993 36.273 2.242 12.261 3.611 24.725 3.506 37.272a.61.61 0 0 0 .545.612.61.61 0 0 0 .672-.541c1.36-12.572 1.239-25.329.205-37.991-1.04-12.672-2.999-25.268-5.461-37.735-2.453-12.468-5.461-24.821-8.472-37.042l-4.232-18.144a115.779 115.779 0 0 1-1.105-6.158c2.079-1.76 4.026-3.877 5.802-6.25 6.739-9.007 11.05-26.939 11.05-36.8 0-17.306-13.272-31.333-29.643-31.333-16.373 0-29.646 14.028-29.646 31.333" class="B C"/></mask><g mask="url(#B)" class="B E"><path d="m146.353 267.495-1.378-4.133-1.224-4.194c-.491-1.355-.731-2.791-1.243-4.12l-1.559-3.949c-.5-1.301-1.457-2.443-1.882-3.675-.321-.945.343-1.164.032-2.112-.192-.637-.669-1.162-.845-1.801-.36-1.269-1.095-2.431-1.373-3.711-.28-1.286.086-2.664-.082-3.966-.151-1.062-1.203-2.022-1.281-3.091-.076-1.062 1.016-2.166 1.008-3.241l-.055-3.169c.055-1.058-.92-2.205-.804-3.279.112-1.051 1.309-1.971 1.476-3.042l.558-4.253.968-4.382c.372-1.566-.359-1.74.018-3.303s1.118-1.386 1.496-2.951-.46-1.767-.086-3.333c.259-1.112 1.366-2.049 1.608-3.197.236-1.123-.396-2.427-.192-3.579l.528-3.418c.116-.886.438-1.741.521-2.637s.425-1.764.462-2.661c.044-.897-.659-1.79-.669-2.687s.287-1.782.225-2.674c-.114-1.783.681-3.694.347-5.402s-2.07-3.018-2.621-4.604c-.563-1.581-.52-3.365-1.28-4.791l-1.055-2.185c-.43-.665-1.319-1.025-1.777-1.657-.343-.475-.631-.25-.984-.719s.112-.809-.244-1.269l-1.401-1.831c-.317-.351-.179-.484-.503-.829-.216-.222-.297-.567-.509-.798-.425-.458-.64-1.073-1.05-1.556-.819-.991-2.115-1.608-2.896-2.672-.761-1.039-1.8-1.916-2.525-3.014-.733-1.113-1.214-2.409-1.921-3.568-.725-1.151-.877-2.645-1.574-3.838-.686-1.205-2.297-1.909-2.98-3.147l-2.069-3.665c-.008-.013 1.136-.637 1.129-.65-.265-.486-1.219.325-1.816.727-.379.257-1.18.081-1.241.454-.265 1.703.053 3.354-.025 4.989-.026.808-.98 1.604-.988 2.411.006.802.999 1.59 1.025 2.389.055 1.596.435 3.117.597 4.687.183 1.561-.427 3.232-.127 4.761.197 1.018 1.462 1.785 1.716 2.786s-.338 2.235-.023 3.221.229 2.149.606 3.115c.296.715.851 1.352 1.181 2.059.16.356.691.539.87.887.261.514-.081.681.172 1.198.543 1.147.887 2.323 1.33 3.486.463 1.152 1.861 1.975 2.201 3.148l.937 3.579c.23 1.202-.457 2.538-.363 3.769l.181 3.651c-.016 1.103.374 2.256.265 3.408-.104 1.097-.62 2.173-.794 3.318-.165 1.09.155 2.296-.061 3.433-.207 1.085-.648 2.155-.89 3.286-.233 1.082-.043 2.287-.3 3.412l-.79 3.206c-.238 1.063-1.636 1.888-1.868 2.98-.227 1.068-.305 2.19-.515 3.285-.325 1.68-.364 3.45-.615 5.179l-.695 5.229-.377 3.484.035 3.506c.057 1.164.948 2.268 1.09 3.405.201 1.689-.32 3.519.065 5.127l1.313 4.782c.349.992.99 1.904 1.404 2.845.423.96 1.31 1.704 1.792 2.613.491.926 1.313 1.638 1.855 2.514.4.646 1.078 1.074 1.493 1.701.44.609.574 1.398 1.037 1.992.233.293-.316 1.243-.08 1.525l.686.855c.472.548 1.514.598 1.997 1.122l1.488 1.523c.486.514.778 1.222 1.283 1.717 1.001.998 2.602 1.352 3.646 2.308 1.042.963 1.498 2.582 2.584 3.508 1.076.916 1.951 2.147 3.1 3.005.663.495 2.356-.149 3.072-.956.492-.552.498-1.374.346-1.899-.027-.051-.112.028-.113.023"/></g><mask id="C" maskUnits="userSpaceOnUse" x="117" y="58" width="67" height="212" mask-type="alpha"><path d="M120.745 89.334c0 12.514 6.944 32.204 16.976 42.156-.302 2.262-.739 4.65-1.228 7.052-1.233 5.953-2.74 12.041-4.232 18.144-3.012 12.221-6.02 24.574-8.472 37.042-2.462 12.467-4.421 25.063-5.46 37.735-1.035 12.662-1.157 25.419.205 37.991a.61.61 0 0 0 .612.545.61.61 0 0 0 .604-.616c-.104-12.547 1.264-25.011 3.505-37.272 2.247-12.266 5.378-24.341 8.993-36.273 3.611-11.943 7.749-23.713 11.959-35.581l6.17-18.115c.005-.014.008-.028.014-.042a.33.33 0 0 0 .012.042c1.949 6.215 4.074 12.169 6.17 18.115 4.21 11.868 8.348 23.638 11.958 35.581 3.617 11.932 6.747 24.007 8.993 36.273 2.242 12.261 3.611 24.725 3.506 37.272a.61.61 0 0 0 .545.612.61.61 0 0 0 .672-.541c1.36-12.572 1.239-25.329.205-37.991-1.04-12.672-2.999-25.268-5.461-37.735-2.453-12.468-5.461-24.821-8.472-37.042l-4.232-18.144a115.779 115.779 0 0 1-1.105-6.158c2.079-1.76 4.026-3.877 5.802-6.25 6.739-9.007 11.05-26.939 11.05-36.8 0-17.306-13.272-31.333-29.643-31.333-16.373 0-29.646 14.028-29.646 31.333" class="B C"/></mask><g mask="url(#C)" class="B E"><path d="m172.551 183.677-.579-4.348c-.262-1.44-1.626-2.657-1.919-4.074-.279-1.414.02-2.93-.283-4.321l-.886-4.149-.694-4.039c-.174-.984-.609-.907-.772-1.892-.092-.661-.468-1.288-.545-1.946-.162-1.31-.118-2.654-.198-3.96l-.087-3.96c.015-1.072.29-2.137.377-3.204s.319-2.109.474-3.173c.154-1.051.79-2.005 1.009-3.061.216-1.037-.079-2.19.199-3.236.271-1.021.19-2.168.52-3.202l1.451-4.075c.5-1.377 1.99-2.404 2.561-3.83.607-1.49.179-1.664.789-3.152s.113-1.694.724-3.183.862-1.387 1.469-2.878c.425-1.059.419-2.297.832-3.394.404-1.074 1.061-2.057 1.438-3.164.371-1.09 1.516-1.951 1.844-3.073.25-.857-.177-1.912.043-2.785s.174-1.788.347-2.668c.179-.881.629-1.702.756-2.59s.812-1.728.886-2.62c.159-1.779.162-3.61.092-5.35s-.925-3.39-1.228-5.041c-.315-1.649-.906-3.164-1.441-4.688-.253-.77-.682-1.428-.98-2.168-.322-.723-.241-1.613-.597-2.307-.266-.522.164-.751-.111-1.267s-.425-.436-.707-.943c-.371-.681-1.009-1.198-1.417-1.847-.259-.397-.505-.227-.773-.618l-.549-.759c-.35-.517.09-1.54-.241-2.08-.659-1.103-2.132-1.766-2.741-2.937-.595-1.142-.944-2.437-1.492-3.634-.556-1.211-1.501-2.275-2.023-3.529-.542-1.247-.097-2.907-.603-4.192l-1.379-3.965-1.567-3.903c-.005-.014-.565.187-.57.172-.189-.521-.382-.189-1.032.119-.415.195-1.28-.124-1.397.235l-1.467 4.804c-.149.795.094 1.674-.037 2.469-.117.793.113 1.63.015 2.423-.187 1.585-.776 3.125-.853 4.701-.057 1.57.102 3.158.166 4.714.041 1.037.945 1.999 1.044 3.029s-1.125 2.233-.963 3.254 1.42 1.842 1.646 2.853l.542 2.251c.104.376-.011.81.114 1.178l.332 1.107 1.034 3.629c.282 1.209 1.079 2.359 1.238 3.57s-.923 2.522-.903 3.739c.045 1.223.26 2.413.166 3.645-.093 1.208.07 2.453-.141 3.681-.185 1.087-.822 2.095-1.107 3.216l-.958 3.229c-.33 1.053-.1 2.317-.487 3.408l-1.016 3.248c-.394 1.033-1.555 1.796-1.979 2.87l-1.198 3.078-1.219 3.082c-.387 1.02.036 2.354-.338 3.403l-1.656 4.898c-.499 1.671-1.771 3.151-2.166 4.86-.251 1.131.59 2.472.4 3.609-.194 1.145-1.032 2.181-1.152 3.324-.121 1.158.302 2.332.27 3.478-.059 1.701-.394 3.363-.258 5.009.118 1.65-.33 3.381-.032 4.955.194 1.033 1.382 1.834 1.649 2.828l.9 2.95c.345.991.027 2.224.429 3.173.297.7 1.615.919 1.929 1.602l.951 1.992c.186.326-.739 1.311-.55 1.626.286.466 1.034.006 1.322.471.381.614.017 1.709.415 2.302s1.623.614 2.034 1.189c.403.582.524 1.362.948 1.927.838 1.139 1.883 2.106 2.77 3.209.884 1.113 1.303 2.615 2.235 3.696.923 1.069 1.976 2.059 2.982 3.084.579.591 1.938.349 2.769-.338.571-.472.868-1.108.798-1.651.111-.189.382-.091.382-.095"/></g><mask id="D" maskUnits="userSpaceOnUse" x="117" y="58" width="67" height="212" mask-type="alpha"><path d="M120.745 89.334c0 12.514 6.944 32.204 16.976 42.156-.302 2.262-.739 4.65-1.228 7.052-1.233 5.953-2.74 12.041-4.232 18.144-3.012 12.221-6.02 24.574-8.472 37.042-2.462 12.467-4.421 25.063-5.46 37.735-1.035 12.662-1.157 25.419.205 37.991a.61.61 0 0 0 .612.545.61.61 0 0 0 .604-.616c-.104-12.547 1.264-25.011 3.505-37.272 2.247-12.266 5.378-24.341 8.993-36.273 3.611-11.943 7.749-23.713 11.959-35.581l6.17-18.115c.005-.014.008-.028.014-.042a.33.33 0 0 0 .012.042c1.949 6.215 4.074 12.169 6.17 18.115 4.21 11.868 8.348 23.638 11.958 35.581 3.617 11.932 6.747 24.007 8.993 36.273 2.242 12.261 3.611 24.725 3.506 37.272a.61.61 0 0 0 .545.612.61.61 0 0 0 .672-.541c1.36-12.572 1.239-25.329.205-37.991-1.04-12.672-2.999-25.268-5.461-37.735-2.453-12.468-5.461-24.821-8.472-37.042l-4.232-18.144a115.779 115.779 0 0 1-1.105-6.158c2.079-1.76 4.026-3.877 5.802-6.25 6.739-9.007 11.05-26.939 11.05-36.8 0-17.306-13.272-31.333-29.643-31.333-16.373 0-29.646 14.028-29.646 31.333" class="B C"/></mask><g mask="url(#D)" class="B F"><path d="M207.673 267.092c-.601-1.34-1.576-2.497-2.228-3.783l-1.801-3.986-2.165-3.71c-.682-1.252-1.532-2.399-2.217-3.623l-2.138-3.519c-.446-.892-.623-.808-1.059-1.705-.277-.606-.067-1.435-.329-2.044-.529-1.211-.493-2.61-.941-3.839-.451-1.237-1.396-2.276-1.739-3.545-.292-1.03-1.209-1.918-1.431-2.965s.831-2.308.678-3.372c-.152-1.052-1.147-1.995-1.236-3.07-.088-1.055.561-2.152.53-3.233-.03-1.056-.809-2.113-.788-3.197.02-1.406.468-2.825.565-4.29.086-1.462-.525-3.024-.382-4.555.157-1.6 1.419-1.474 1.581-3.073s-.974-1.717-.812-3.32l.198-3.215c.105-1.136.056-2.286.14-3.456.082-1.143 1.176-2.238 1.221-3.407l.27-3.493c-.005-.893-.286-1.799-.324-2.698s.296-1.84.21-2.733c-.078-.895-.806-1.732-.937-2.618s-.048-1.825-.231-2.701c-.354-1.751-1.027-3.42-1.589-5.068s-1.348-3.175-2.109-4.672c-.772-1.491-1.959-2.658-2.906-3.969-.461-.666-.696-1.465-1.192-2.089-.515-.602-.414-1.775-.954-2.339l-.793-.882c-.409-.413-.871.044-1.286-.363-.548-.548-1.188-.984-1.763-1.491-.361-.305-.448-.198-.817-.496-.244-.189-.154-.789-.395-.99-.484-.396-1.248-.467-1.718-.891-.946-.871-1.431-2.222-2.348-3.17-.894-.928-2.677-1.064-3.544-2.056-.878-1.002-1.338-2.391-2.195-3.444-.874-1.041-1.26-2.479-2.111-3.567-.844-1.102-2.417-1.672-3.26-2.806l-2.605-3.306c-.009-.013.396-.313.387-.325-.328-.446-.795-.337-1.332.144-.341.306.059.589.05.967l.034 4.9c.083.805-.662 1.697-.56 2.498.114.793.431 1.558.564 2.348.271 1.573 1.384 2.925 1.759 4.46l1.166 4.497c.333.983.407 2.039.795 2.996s.93 1.836 1.377 2.77.708 1.947 1.213 2.852c.39.67.143 1.742.565 2.397.207.331 1.129.195 1.354.514.328.474-.08.752.24 1.232l2.048 3.136c.614 1.079.727 2.357 1.224 3.473s1.818 1.885 2.184 3.045c.39 1.159.173 2.482.432 3.69.255 1.185 1.356 2.256 1.504 3.494.132 1.096-.535 2.3-.487 3.456.044 1.101.511 2.228.494 3.386l-.194 3.396c-.058 1.102-.521 2.207-.608 3.36-.083 1.103.358 2.275.258 3.426-.094 1.079-.365 2.16-.46 3.273-.093 1.085-.742 2.143-.825 3.256-.08 1.089-.461 2.181-.521 3.294-.095 1.709.725 3.479.709 5.225l.247 5.23c.083 1.155-.345 2.357-.204 3.501l.306 3.483c.213 1.145 1.018 2.165 1.312 3.273.428 1.647 1.384 3.07 1.983 4.609.583 1.549.595 3.302 1.329 4.726.479.934 1.375 1.615 1.914 2.49l1.692 2.589c.611.85 1.286 1.613 1.942 2.408.483.587.693 1.392 1.188 1.958.518.545 1.44.665 1.979 1.19l.778.785c.405.365.545.206.953.57.541.479.475 1.637 1.025 2.091s1.653.212 2.209.645c.553.445.59 1.543 1.157 1.964 1.127.854 2.558 1.27 3.723 2.076 1.163.814 1.943 2.204 3.144 2.974 1.19.762 2.86.723 4.115 1.419.723.401 1.024.157 1.625-.738.413-.614.777-1.022.555-1.522.053-.212.611-.129.608-.132"/></g><mask id="E" maskUnits="userSpaceOnUse" x="117" y="58" width="67" height="212" mask-type="alpha"><path d="M120.745 89.334c0 12.514 6.944 32.204 16.976 42.156-.302 2.262-.739 4.65-1.228 7.052-1.233 5.953-2.74 12.041-4.232 18.144-3.012 12.221-6.02 24.574-8.472 37.042-2.462 12.467-4.421 25.063-5.46 37.735-1.035 12.662-1.157 25.419.205 37.991a.61.61 0 0 0 .612.545.61.61 0 0 0 .604-.616c-.104-12.547 1.264-25.011 3.505-37.272 2.247-12.266 5.378-24.341 8.993-36.273 3.611-11.943 7.749-23.713 11.959-35.581l6.17-18.115c.005-.014.008-.028.014-.042a.33.33 0 0 0 .012.042c1.949 6.215 4.074 12.169 6.17 18.115 4.21 11.868 8.348 23.638 11.958 35.581 3.617 11.932 6.747 24.007 8.993 36.273 2.242 12.261 3.611 24.725 3.506 37.272a.61.61 0 0 0 .545.612.61.61 0 0 0 .672-.541c1.36-12.572 1.239-25.329.205-37.991-1.04-12.672-2.999-25.268-5.461-37.735-2.453-12.468-5.461-24.821-8.472-37.042l-4.232-18.144a115.779 115.779 0 0 1-1.105-6.158c2.079-1.76 4.026-3.877 5.802-6.25 6.739-9.007 11.05-26.939 11.05-36.8 0-17.306-13.272-31.333-29.643-31.333-16.373 0-29.646 14.028-29.646 31.333" class="B C"/></mask><g mask="url(#E)" class="B F"><path d="M178.72 181.51c-.166-1.459.047-2.953-.183-4.378-.233-1.444-.161-2.938-.426-4.359-.25-1.42-.768-2.778-1.044-4.175-.27-1.401-1.54-2.589-1.821-3.962-.269-1.369.619-2.964.411-4.251-.155-.986-1.069-.846-1.213-1.834-.079-.662.755-1.431.691-2.09-.136-1.314-1.365-2.547-1.419-3.854-.054-1.316.724-2.65.781-3.961.035-1.072.224-2.116.331-3.181s-.231-2.182-.054-3.243l.537-3.178c.237-1.033.207-2.17.505-3.211.291-1.016 1.714-1.694 2.063-2.72.446-1.333 1.075-2.603 1.611-3.969.525-1.367.788-2.862 1.388-4.277.636-1.477-.228-1.851.411-3.326s1.42-1.14 2.061-2.616l1.336-2.933c.445-1.048.544-2.266.979-3.355.426-1.066 1.455-1.924 1.854-3.025.392-1.082-.204-2.512.145-3.626.267-.853.98-1.572 1.216-2.441s.874-1.646 1.064-2.522c.197-.878-.75-1.97-.607-2.854s1.228-1.639 1.318-2.529c.195-1.775-.398-3.568-.434-5.309s.578-3.526.308-5.183c-.283-1.654-1.571-3.04-2.075-4.575-.238-.775.566-1.921.282-2.666-.308-.73-.852-1.398-1.194-2.101-.257-.527-.457-.434-.723-.955s-.089-.613-.362-1.126c-.357-.689-1.729-.764-2.123-1.422-.252-.401-.201-.417-.462-.812-.174-.255.842-1.29.674-1.554-.34-.524-.918-.877-1.238-1.424-.637-1.116-1.499-2.114-2.085-3.296-.572-1.154-.932-2.411-1.457-3.619-.531-1.221-1.448-2.293-1.944-3.557l-1.711-3.753c-.469-1.306-.137-2.912-.597-4.249-.453-1.319-1.989-2.279-2.44-3.612-.004-.014 1.141-.401 1.135-.416-.177-.526-1.456-.776-2.112-.482-.419.188-.026.971-.15 1.327-.552 1.634-2.067 2.883-2.424 4.479l-.616 2.374c-.132.79.728 1.741.615 2.532-.218 1.582-.279 3.132-.387 4.708l-.232 4.665c.019 1.037-.309 2.088-.231 3.118s.907 1.974 1.048 2.999l.413 3.044c.169.756.489 1.452.692 2.204.096.379-.325.889-.207 1.261.168.552-.183.655-.023 1.208.337 1.223.397 2.464.633 3.686.258 1.215 1.05 2.314 1.185 3.528s.175 2.427.17 3.644c.021 1.223-.231 2.424-.35 3.653-.117 1.207.656 2.579.423 3.802-.207 1.084-.546 2.221-.851 3.338-.292 1.061-1.596 1.855-1.964 2.952-.35 1.047-1.1 1.957-1.509 3.041-.39 1.032.084 2.425-.348 3.498l-1.18 3.187c-.416.998-1.612 1.694-2.04 2.727l-1.386 3.016-1.097 3.137-1.687 4.936c-.533 1.66.009 3.637-.42 5.337-.272 1.126-1.039 2.125-1.253 3.258l-.732 3.398c-.145 1.156-.652 2.298-.708 3.443-.092 1.698.928 3.384 1.031 5.032.086 1.653-.224 3.303.043 4.882.173 1.036.149 2.113.396 3.111.253 1.017 1.331 1.778 1.65 2.756l.867 2.966c.283.705.034 1.629.336 2.319.326.677 1.152 1.093 1.507 1.758.179.329.164.768.347 1.086.276.471.556.306.833.776.37.62.452 1.399.839 1.999s.447 1.412.846 1.995c.391.589.986 1.043 1.398 1.618l2.318 3.564c.862 1.129 2.459 1.666 3.37 2.764.902 1.087 1.046 2.849 2.032 3.893.567.601 2.354.663 3.198-.009.579-.46.98-1.766.919-2.311-.065-.107-.691.286-.691.282"/></g><mask id="F" maskUnits="userSpaceOnUse" x="117" y="58" width="67" height="212" mask-type="alpha"><path d="M120.745 89.334c0 12.514 6.944 32.204 16.976 42.156-.302 2.262-.739 4.65-1.228 7.052-1.233 5.953-2.74 12.041-4.232 18.144-3.012 12.221-6.02 24.574-8.472 37.042-2.462 12.467-4.421 25.063-5.46 37.735-1.035 12.662-1.157 25.419.205 37.991a.61.61 0 0 0 .612.545.61.61 0 0 0 .604-.616c-.104-12.547 1.264-25.011 3.505-37.272 2.247-12.266 5.378-24.341 8.993-36.273 3.611-11.943 7.749-23.713 11.959-35.581l6.17-18.115c.005-.014.008-.028.014-.042a.33.33 0 0 0 .012.042c1.949 6.215 4.074 12.169 6.17 18.115 4.21 11.868 8.348 23.638 11.958 35.581 3.617 11.932 6.747 24.007 8.993 36.273 2.242 12.261 3.611 24.725 3.506 37.272a.61.61 0 0 0 .545.612.61.61 0 0 0 .672-.541c1.36-12.572 1.239-25.329.205-37.991-1.04-12.672-2.999-25.268-5.461-37.735-2.453-12.468-5.461-24.821-8.472-37.042l-4.232-18.144a115.779 115.779 0 0 1-1.105-6.158c2.079-1.76 4.026-3.877 5.802-6.25 6.739-9.007 11.05-26.939 11.05-36.8 0-17.306-13.272-31.333-29.643-31.333-16.373 0-29.646 14.028-29.646 31.333" class="B C"/></mask><g mask="url(#F)" class="B F"><path d="m118.978 95.31.536 4.305c.232 1.445-.496 3.044-.231 4.466.251 1.419.739 2.784 1.014 4.18l.999 4.129.843 4.016c.154.985-.438 1.07-.295 2.057.08.662 1.106 1.193 1.169 1.852.137 1.314.082 2.645.136 3.953l.021 3.95-.246 3.203c-.106 1.062-.608 2.067-.784 3.126-.174 1.049.054 2.179-.187 3.23l-.726 3.138c-.292 1.016-.747 2.002-1.096 3.028-.446 1.334-1.489 2.475-2.025 3.839-.526 1.368-1.305 2.647-1.905 4.063-.636 1.476-.288 1.625-.928 3.1s-.816 1.401-1.456 2.878l-1.38 2.915c-.445 1.05.261 2.597-.174 3.686-.426 1.065-1.889 1.748-2.288 2.849-.391 1.082-.066 2.422-.415 3.537l-.854 2.547c-.236.864-1.317 1.536-1.507 2.414-.198.877.108 1.878-.037 2.764l-.553 2.662c-.194 1.775.61 3.581.646 5.321s.279 3.373.55 5.029c.282 1.655-.09 3.431.414 4.966l.712 2.326c.308.729 1.453 1.084 1.795 1.785.256.527-.524.912-.257 1.433s.936.168 1.208.68l1.07 2.013c.252.401-.033.58.227.975.176.255.827.205.995.469.341.524.429 1.236.749 1.783.636 1.117.902 2.461 1.487 3.643.572 1.154 1.628 2.11 2.154 3.318l1.587 3.724c.517 1.258-.137 2.972.343 4.266.469 1.305 2.013 2.245 2.473 3.583l1.463 3.946c.005.015-.595.218-.59.233.177.525 1.041.985 1.698.691.418-.188 1.212-.618 1.335-.975.553-1.632-.002-3.482.355-5.077.165-.793 1.35-1.389 1.496-2.184.132-.79.143-1.62.255-2.411.218-1.581-.957-3.278-.848-4.853.088-1.569.775-3.091.741-4.647l.027-3.113c-.079-1.033-.746-1.994-.887-3.019s-.502-1.974-.708-2.991c-.169-.756.464-1.695.261-2.447-.096-.379-.933-.537-1.05-.909-.169-.551.468-.746.308-1.298-.338-1.223-1.017-2.354-1.253-3.576-.258-1.215-.69-2.429-.824-3.643s.839-2.481.844-3.697c-.021-1.223.113-2.376.232-3.606.117-1.205-.317-2.471-.082-3.694.205-1.085-.083-2.327.223-3.445.29-1.062.733-2.134 1.101-3.233.349-1.045.959-2.039 1.368-3.122.389-1.033.379-2.26.812-3.333.414-1.026 1.399-1.848 1.844-2.914.417-.999 1.077-1.917 1.506-2.949.417-1.006.301-2.246.717-3.282.408-1.013 1.288-1.853 1.683-2.895.61-1.6.593-3.435 1.137-5.094l1.531-4.997c.272-1.124 1.195-2.108 1.409-3.241.215-1.142.732-2.264.875-3.405.144-1.155-1.04-2.424-.984-3.569.092-1.698.531-3.339.427-4.989-.086-1.652-.672-3.184-.938-4.763-.174-1.036.112-2.129-.136-3.128l-.871-2.982c-.325-.996-.479-2.045-.863-3-.282-.705-.083-1.634-.385-2.323-.326-.677-1.388-.997-1.742-1.661-.18-.329-.465-.6-.647-.919-.276-.471.115-.7-.163-1.172-.37-.62-.3-1.523-.687-2.123s-1.851-.487-2.25-1.069c-.392-.591-.547-1.33-.96-1.904-.816-1.155-1.806-2.156-2.671-3.278l-2.655-3.316c-.902-1.088-1.199-2.713-2.184-3.756-.568-.602-2.204-.647-3.048.025-.58.461-.762 1.601-.702 2.146.214.172.696-.246.696-.241"/></g><g class="B"><use xlink:href="#H" class="D"/><use xlink:href="#I" fill="#1e0135"/><use xlink:href="#H" x="-34.814" class="D"/><use xlink:href="#I" x="-34.814" fill="#1e0135"/></g><defs><path id="H" d="M177.644 122.839a9.85 9.85 0 0 1-9.846 9.847 9.85 9.85 0 0 1-9.848-9.847 9.85 9.85 0 0 1 9.848-9.846 9.85 9.85 0 0 1 9.846 9.846"/><path id="I" d="M173.638 122.839a5.84 5.84 0 0 1-5.841 5.841 5.841 5.841 0 0 1-5.84-5.841 5.84 5.84 0 1 1 11.681 0"/></defs></svg>',
   '<svg width="300" height="300" fill="none" xmlns="http://www.w3.org/2000/svg"><rect width="300" height="300" rx="150" fill="#FFFBF8"/><path d="m143.145 128.265-29.01 16.034a3.85 3.85 0 0 1-2.922.329 3.861 3.861 0 0 1-1.332-.688 3.859 3.859 0 0 1-.967-1.145l-10.325-18.682a3.842 3.842 0 0 1 1.504-5.221l17.24-9.529a17.29 17.29 0 0 1 23.495 6.768l3.806 6.886a3.842 3.842 0 0 1-1.504 5.221l.015.027ZM116.913 191.176l-17.24-9.529a3.843 3.843 0 0 1-1.504-5.221l10.326-18.681a3.835 3.835 0 0 1 3.793-1.959c.501.056.986.211 1.428.455l29.009 16.033a3.853 3.853 0 0 1 1.834 2.299 3.847 3.847 0 0 1-.33 2.923l-3.805 6.885a17.292 17.292 0 0 1-23.511 6.795Z" fill="#1E0135"/><path d="M90.433 150.362a99.28 99.28 0 0 1 79.784-40.192A99.278 99.278 0 0 1 250 150.362a99.285 99.285 0 0 1-159.567 0Z" fill="#A0F"/><path d="M217.952 140.788a4.727 4.727 0 1 0 0-9.453 4.727 4.727 0 0 0 0 9.453Z" fill="#1E0135"/><path fill-rule="evenodd" clip-rule="evenodd" d="M191.668 150.057a7.715 7.715 0 1 0 0-15.429M191.668 168.17a7.714 7.714 0 1 0 0-15.429" fill="#FFFF8D"/><path fill-rule="evenodd" clip-rule="evenodd" d="M141.05 150.057a7.714 7.714 0 0 0 0-15.429M141.05 168.17a7.713 7.713 0 1 0 0-15.429" fill="#EA80FC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M124.187 158.077a7.714 7.714 0 0 0 0-15.43" fill="#FFFF8D"/><path fill-rule="evenodd" clip-rule="evenodd" d="M174.805 158.412a7.716 7.716 0 0 0 0-15.246M174.805 139.628a7.713 7.713 0 1 0 0-15.428M174.805 176.525a7.713 7.713 0 0 0 0-15.399" fill="#EA80FC"/><path fill-rule="evenodd" clip-rule="evenodd" d="M157.912 158.413a7.723 7.723 0 0 0 8.193-4.399 7.72 7.72 0 0 0 0-6.45 7.701 7.701 0 0 0-1.99-2.633 7.707 7.707 0 0 0-6.203-1.765M157.912 139.628a7.713 7.713 0 1 0 0-15.428M157.912 176.525a7.702 7.702 0 0 0 5.766-2.075 7.706 7.706 0 0 0 1.799-8.688 7.717 7.717 0 0 0-7.565-4.635" fill="#FFFF8D"/><path d="m56.311 178.262 58.424-27.901-58.424-27.9a64.824 64.824 0 0 0 0 55.801Z" fill="#A0F"/></svg>',
   '<svg width="300" height="300" fill="none" xmlns="http://www.w3.org/2000/svg"><rect width="300" height="300" rx="150" fill="#C6FF00"/><path fill-rule="evenodd" clip-rule="evenodd" d="m241.017 201.789.028-3.038c-8.275-.071-18.039-1.037-28.637-2.874 6.231-5.998 8.505-15.859 4.185-24.465l17.85 8.723c9.19-19.842 9.364-39.763.519-59.762l-18.717 8.745c4.946-8.463 3.024-18.619-3.135-24.855 10.666-1.663 20.481-2.46 28.777-2.389l.028-3.037-42.695-.373-16.116-.14-37.031-.322c-23.301-.203-43.184 15.591-49.182 37.07-12.12 12.026-24.744 16.36-38.805 16.238l-.004.34-.08 9.358c13.695.121 32.805 4.509 45.501 16.048 9.275 14.234 25.268 23.755 43.433 23.913l35.272.309 58.809.511Z" fill="#FF80AB"/><mask id="a" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="58" y="98" width="184" height="104"><path fill-rule="evenodd" clip-rule="evenodd" d="M96.887 135.085c-12.12 12.02-24.744 16.348-38.803 16.225l-.002.34-.082 9.359c13.682.12 32.776 4.488 45.47 15.998 9.271 14.262 25.279 23.803 43.465 23.964l35.271.307 58.811.512.026-3.039c-8.273-.071-18.037-1.036-28.635-2.875 6.229-5.995 8.505-15.858 4.183-24.463l17.85 8.722c9.19-19.841 9.364-39.761.521-59.761l-18.719 8.744c4.948-8.463 3.024-18.618-3.133-24.854 10.666-1.664 20.479-2.461 28.777-2.391l.024-3.037-42.693-.37-16.114-.142-37.033-.322-.443-.001c-23.115-.002-42.785 15.735-48.74 37.084Z" fill="#fff"/></mask><g mask="url(#a)"><path fill-rule="evenodd" clip-rule="evenodd" d="M100.7 105.421s6.965 13.16 4.431 58.573c-1.486 26.672 1.668 34.078-.407 37.153-2.077 3.077 21.598 6.561 23.236 6.607 1.638.045 7.804-1.92 5.437-4.748-2.366-2.83-4.768-39.353-4.9-64.1-.133-24.746 6.889-46.545 4.856-48.47-2.033-1.927-17.693-7.79-17.693-7.79l-14.96 22.775Z" fill="#1E0135"/></g><mask id="b" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="58" y="98" width="184" height="104"><path fill-rule="evenodd" clip-rule="evenodd" d="M96.887 135.085c-12.12 12.02-24.744 16.348-38.803 16.225l-.002.34-.082 9.359c13.682.12 32.776 4.488 45.47 15.998 9.271 14.262 25.279 23.803 43.465 23.964l35.271.307 58.811.512.026-3.039c-8.273-.071-18.037-1.036-28.635-2.875 6.229-5.995 8.505-15.858 4.183-24.463l17.85 8.722c9.19-19.841 9.364-39.761.521-59.761l-18.719 8.744c4.948-8.463 3.024-18.618-3.133-24.854 10.666-1.664 20.479-2.461 28.777-2.391l.024-3.037-42.693-.37-16.114-.142-37.033-.322-.443-.001c-23.115-.002-42.785 15.735-48.74 37.084Z" fill="#fff"/></mask><g mask="url(#b)"><path fill-rule="evenodd" clip-rule="evenodd" d="M202.03 82.975s11.068 20.912 7.044 93.085c-2.364 42.388 2.647 54.156-.651 59.045-3.3 4.89 34.324 10.426 36.927 10.499 2.605.071 12.403-3.051 8.642-7.547-3.76-4.495-7.578-62.54-7.788-101.867-.21-39.325 10.947-73.97 7.719-77.03-3.231-3.063-28.12-12.381-28.12-12.381L202.03 82.975Z" fill="#1E0135"/></g><mask id="c" style="mask-type:alpha" maskUnits="userSpaceOnUse" x="58" y="98" width="184" height="104"><path fill-rule="evenodd" clip-rule="evenodd" d="M96.887 135.085c-12.12 12.02-24.744 16.348-38.803 16.225l-.002.34-.082 9.359c13.682.12 32.776 4.488 45.47 15.998 9.271 14.262 25.279 23.803 43.465 23.964l35.271.307 58.811.512.026-3.039c-8.273-.071-18.037-1.036-28.635-2.875 6.229-5.995 8.505-15.858 4.183-24.463l17.85 8.722c9.19-19.841 9.364-39.761.521-59.761l-18.719 8.744c4.948-8.463 3.024-18.618-3.133-24.854 10.666-1.664 20.479-2.461 28.777-2.391l.024-3.037-42.693-.37-16.114-.142-37.033-.322-.443-.001c-23.115-.002-42.785 15.735-48.74 37.084Z" fill="#fff"/></mask><g mask="url(#c)"><path fill-rule="evenodd" clip-rule="evenodd" d="M150.548 92.34s9.34 17.652 5.943 78.57c-1.992 35.776 2.238 45.708-.548 49.835-2.785 4.126 28.971 8.8 31.17 8.861 2.198.06 10.468-2.574 7.295-6.368-3.176-3.797-6.397-52.788-6.574-85.983-.178-33.192 9.239-62.434 6.513-65.017-2.725-2.585-23.733-10.45-23.733-10.45l-20.066 30.553Z" fill="#1E0135"/></g><path fill-rule="evenodd" clip-rule="evenodd" d="M121.66 127.308a3.864 3.864 0 1 1-7.73-.066 3.864 3.864 0 0 1 7.73.066" fill="#18FFFF"/><path fill-rule="evenodd" clip-rule="evenodd" d="M119.838 127.291a2.041 2.041 0 0 1-2.061 2.024 2.038 2.038 0 0 1-2.023-2.059 2.043 2.043 0 0 1 4.084.035" fill="#1E0135"/></svg>',
@@ -176,6 +177,23 @@ interface TokenIcon extends BaseIcon {
   src: string;
   address?: string
 }
+
+/**
+ *  Returns icon url if exists, otherwise return sample icon based on calculated
+ *  checksum from provided address. Returned sample icon is svg encoded to base64
+ *  and prefixed with data string, so it can be used directly with <img /> tag.
+ */
+export const getIconUrl = (tokenAddress = ''): string => {
+  const reefToken = reefTokenWithAmount();
+
+  if (tokenAddress === reefToken.address) {
+    return reefToken.iconUrl;
+  }
+
+  const checkSum = getHashSumLastNr(tokenAddress);
+  const nr = checkSum > -1 && checkSum < 10 ? checkSum : checkSum % 10;
+  return `data:image/svg+xml;base64,${btoa(DEFAULT_TOKEN_ICONS[nr])}`;
+};
 
 export const TokenIcon = ({ src, address }: TokenIcon): JSX.Element => (
   <div className="d-flex my-auto user-select-none">
